// modules/apps/ai-hub-cell/ai-hub-cell-web/src/main/resources/META-INF/resources/js/types/Template.ts
/**
 * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */

export type Template = {
	entries: number;
	icon: string;
	labels: Array<{
		text: string;
		type: 'danger' | 'info' | 'secondary' | 'success' | 'warning';
	}>;
	name: string;
};